Move RefreshableAwsCredentialsProvider to aws package
Not KMS-specific — it caches and refreshes any AwsSessionCredentials
based on expiration time, regardless of how they were obtained.

Issue: world-federation-of-advertisers/cross-media-measurement#3907
