+++ src/main/kotlin/org/wfanet/measurement/aws/BUILD.bazel
@@ -1,0 +1,11 @@
+load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
+
+package(default_visibility = ["//visibility:public"])
+
+kt_jvm_library(
+    name = "refreshable_aws_credentials_provider",
+    srcs = ["RefreshableAwsCredentialsProvider.kt"],
+    deps = [
+        "//imports/java/software/amazon/awssdk/auth",
+    ],
+)
