+++ src/main/kotlin/org/wfanet/measurement/aws/RefreshableAwsCredentialsProvider.kt
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.aws.kms
+package org.wfanet.measurement.aws
 
 import java.time.Clock
 import java.time.Duration
