+++ src/main/kotlin/org/wfanet/measurement/gcloud/kms/BUILD.bazel
@@ -22,7 +22,7 @@ kt_jvm_library(
         "//imports/java/software/amazon/awssdk/services/regions",
         "//imports/java/software/amazon/awssdk/services/sts",
         "//src/main/kotlin/org/wfanet/measurement/aws/kms:aws_kms_client_factory",
-        "//src/main/kotlin/org/wfanet/measurement/aws/kms:refreshable_aws_credentials_provider",
+        "//src/main/kotlin/org/wfanet/measurement/aws:refreshable_aws_credentials_provider",
         "//src/main/kotlin/org/wfanet/measurement/common/crypto/tink",
         "@maven//:com_google_auth_google_auth_library_credentials",
         "@maven//:com_google_auth_google_auth_library_oauth2_http",
