+++ src/test/kotlin/org/wfanet/measurement/aws/BUILD.bazel
@@ -1,0 +1,15 @@
+load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_test")
+
+kt_jvm_test(
+    name = "RefreshableAwsCredentialsProviderTest",
+    srcs = ["RefreshableAwsCredentialsProviderTest.kt"],
+    test_class = "org.wfanet.measurement.aws.RefreshableAwsCredentialsProviderTest",
+    deps = [
+        "//imports/java/com/google/common/truth",
+        "//imports/java/org/junit",
+        "//imports/java/software/amazon/awssdk/auth",
+        "//imports/kotlin/kotlin/test",
+        "//imports/kotlin/kotlinx/coroutines/test",
+        "//src/main/kotlin/org/wfanet/measurement/aws:refreshable_aws_credentials_provider",
+    ],
+)
