+++ src/test/kotlin/org/wfanet/measurement/aws/kms/BUILD.bazel
@@ -12,17 +12,3 @@ kt_jvm_test(
         "//src/main/kotlin/org/wfanet/measurement/aws/kms:aws_kms_client_factory",
     ],
 )
-
-kt_jvm_test(
-    name = "RefreshableAwsCredentialsProviderTest",
-    srcs = ["RefreshableAwsCredentialsProviderTest.kt"],
-    test_class = "org.wfanet.measurement.aws.kms.RefreshableAwsCredentialsProviderTest",
-    deps = [
-        "//imports/java/com/google/common/truth",
-        "//imports/java/org/junit",
-        "//imports/java/software/amazon/awssdk/auth",
-        "//imports/kotlin/kotlin/test",
-        "//imports/kotlin/kotlinx/coroutines/test",
-        "//src/main/kotlin/org/wfanet/measurement/aws/kms:refreshable_aws_credentials_provider",
-    ],
-)
